precisa arrumar ainda

+++ src/euclides.h
@@ -2,6 +2,6 @@
 #define EUCLIDES
 
 // Algoritmo de euclides estendido que salva o MDC entre (a,b) e seus coeficientes na matriz A.
-void euclides(__uint128_t, __uint128_t, __uint128_t A[2][3], long int B[1][3]);
+void euclides(__uint128_t, __uint128_t, __uint128_t A[2][3], unsigned long long int B[1][3]);
 
 #endif
